# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Crypto API-документация",
        "description": "Crypto API-документация",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/v1/"
        }
    ],
    "paths": {
        "/ping": {
            "get": {
                "tags": [
                    "Ping"
                ],
                "summary": "Ping",
                "description": "Ping",
                "operationId": "Ping",
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/create": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Wallet create",
                "description": "Wallet create",
                "operationId": "Wallet create",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "currency"
                                ],
                                "properties": {
                                    "currency": {
                                        "description": "Currency",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/list": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Wallet list",
                "description": "Wallet list",
                "operationId": "Wallet list",
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}